UpdateScore done

+++ app/src/main/java/com/cuongngo/core_project/services/remote/UserRemoteDataSource.kt
@@ -27,8 +27,8 @@ class UserRemoteDataSource(private val apiService: THPApi) : BaseRemoteDataSourc
      suspend fun updateScore(
          examinerId: Int?,
          examId: Int?,
-         score: Float?,
-         composingScore: Float?
+         score: Double?,
+         composingScore: Double?
      ) = getResult {
         apiService.updateScore(
             examinerId, examId, score, composingScore
